+++ timeline/urls.py
@@ -1,0 +1,26 @@
+from django.conf.urls import url
+
+from . import views
+
+
+app_name = 'timeline'
+
+urlpatterns = [
+    url(
+        regex=r'^(?P<project>[0-9a-f-]+)$',
+        view=views.ProjectTimelineView.as_view(),
+        name='project_timeline',
+    ),
+    url(    # TODO: Refactor this to use the object ref model uuid?
+        regex=r'^(?P<project>[0-9a-f-]+)/(?P<object_model>[\w-]+)/'
+              r'(?P<object_uuid>[0-9a-f-]+)$',
+        view=views.ObjectTimelineView.as_view(),
+        name='object_timeline',
+    ),
+    # Taskflow API views
+    url(
+        regex=r'^taskflow/status/set$',
+        view=views.TimelineEventStatusSetAPIView.as_view(),
+        name='taskflow_status_set',
+    ),
+]